pyproject.toml
[tool.poetry]
name = "WP6_QNLP"
version = "0.3.0"
description = "The repository of the quantum natural language processing WP6 within NEASQC. Development and releases are stored in this repository."
authors = [
    "Richard Wolf <richard.wolf@ichec.ie>",
    "Conor Dune <conor.dunne@ichec.ie>",
    "Pablo Suarez <pablo.suarez@ichec.ie>"]
readme = "README.md"
repository = "https://github.com/NEASQC/WP6_QNLP"
packages = [
     {include = "module", from = "./neasqc_wp61/models/quantum/alpha/"}
    ]


[tool.poetry.dependencies]
python = "~3.10"
pandas = "==1.5.1"
numpy = "==1.23.4"
matplotlib = "==3.6.1"
myqlm = "==1.5.1"
scipy = "==1.9.3"
scikit-learn = "==1.1.3"
discopy = "==1.1.4"
lambeq = "0.3.3"
pytket = ">=0.19.2"
qiskit = "==0.43.1"
qiskit_ignis = "==0.7.1"
sympy = "==1.11.1"
torch = "2.0.1"
tensorflow = {version = "==2.10.0", markers = "platform_machine != 'arm64'"}
tensorflow-macos = {version = "==2.10.0", markers = "sys_platform == 'darwin' and platform_machine == 'arm64'"}
tensorflow-metal = {version = "==0.6.0", markers = "sys_platform == 'darwin' and platform_machine == 'arm64'"}
GitPython = "==3.1.31"
nltk = "^3.8.1"
torchvision = "0.15.2"
pennylane = "^0.30.0"
plotly = "^5.15.0"
bayesian-optimization = "^1.4.3"


[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx  = "==4.2.0"
sphinx-rtd-theme =  "==1.0.0"
myst-parser = "^0.18.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.23.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"